Init: Create flutter app (#3)
This pull request includes several changes to the `app/ui` directory,
primarily focusing on adding configuration files and documentation for a
new Flutter project. The most important changes include the addition of
a `.gitignore` file, a `.metadata` file for tracking project properties,
and a `README.md` file to provide an overview and resources for getting
started with Flutter.

Configuration files:

*
[`app/ui/.gitignore`](diffhunk://#diff-9bcbb4ad5739db6eb2607f4f5e3d5455a9adcb577e1082d0ecd7370423c3ab59R1-R45):
Added a `.gitignore` file to exclude various temporary and build-related
files from version control.
*
[`app/ui/.metadata`](diffhunk://#diff-8c4c1fb6ae88cbde95b9c43f078a8afa592a1bd05619ce443abc3a07e8e2cb42R1-R30):
Added a `.metadata` file to track properties of the Flutter project,
including version and migration information.

Documentation:

*
[`app/ui/README.md`](diffhunk://#diff-2c9a39632e644a0e0dc8de59545414b34fbcc4055dab176c4bd89add3f48cc35R1-R16):
Added a `README.md` file to provide an overview of the Flutter project
and links to resources for getting started with Flutter development.