Feat: Add home page (#5)
This pull request includes significant updates to the UI, theme, and
structure of the Ntango app. The changes focus on implementing a new
theme, restructuring the home page, and improving the task card widget.

### UI and Theme Updates:
*
[`app/ui/lib/main.dart`](diffhunk://#diff-6a6538cd8b6a9caa6774f61089449f2b0db4c35e28f8955c2d6a9a170e57a363L2-R3):
Updated the main app to use new light and dark themes
(`ntangoLightTheme` and `ntangoDarkTheme`), and set `HomePage` as the
home screen.
[[1]](diffhunk://#diff-6a6538cd8b6a9caa6774f61089449f2b0db4c35e28f8955c2d6a9a170e57a363L2-R3)
[[2]](diffhunk://#diff-6a6538cd8b6a9caa6774f61089449f2b0db4c35e28f8955c2d6a9a170e57a363L15-R20)
*
[`app/ui/lib/src/theme/ntango_theme.dart`](diffhunk://#diff-c94db56bf1b24ed2ef22ab90484c2b9df46672e0646e05325ea4616ba69b3936R1-R41):
Added new light and dark themes with customized color schemes.

### Home Page and Navigation:
*
[`app/ui/lib/src/presentation/home_page.dart`](diffhunk://#diff-57935ece6da0c2c102dc919232597fca74558836387aabb7b0690239e9c0b673R1-R91):
Created a new `HomePage` with a bottom navigation bar and an indexed
stack to switch between the task list and a placeholder settings page.
*
[`app/ui/pubspec.yaml`](diffhunk://#diff-8c7a5b3868c526d1b4936992a4daed7544c0fd40fb3053f9c156985f157f8decR55-R57):
Updated assets section to include images for the app logo.

### Task Management:
*
[`app/ui/lib/src/features/tasks/presentation/widgets/tasks_list.dart`](diffhunk://#diff-31fb0d646ac51c805738721d6d86606d8092bd7fb8ab47ab9c2e227fe901f57aR1-R30):
Added a new `TasksList` widget to display a list of tasks using
`FutureBuilder`.
*
[`app/ui/lib/src/features/tasks/presentation/widgets/task_card.dart`](diffhunk://#diff-29433baee83d888503eedb15ed1a38e2d3c0044fe198917b1222c87e71193c0dL7-R7):
Enhanced `TaskCard` to use the new theme colors and improved text
styles.
[[1]](diffhunk://#diff-29433baee83d888503eedb15ed1a38e2d3c0044fe198917b1222c87e71193c0dL7-R7)
[[2]](diffhunk://#diff-29433baee83d888503eedb15ed1a38e2d3c0044fe198917b1222c87e71193c0dR32-R55)
[[3]](diffhunk://#diff-29433baee83d888503eedb15ed1a38e2d3c0044fe198917b1222c87e71193c0dL63-R116)

### Code Cleanup:
*
[`app/ui/lib/src/features/tasks/presentation/pages/tasks_page.dart`](diffhunk://#diff-5cfbe2b88af34500d04ecd8a274b48d752066228d885bf7e78692efb1e498569L1-L40):
Removed the old `TasksPage` class as it has been replaced by the new
`TasksList` widget.